requirements.txt
Babel==2.7.0
dj-database-url==0.5.0
Django==1.11.23
django-autoslug==1.9.6
django-bootstrap4==0.0.7
django-categories==1.6.1
django-crispy-forms==1.8.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-money==0.15.1
django-mptt==0.9.1
django-phone-field==1.8.0
django-phonenumber-field==3.0.1
django-search-views==0.3.7
django-simple-search==1.0.2
django-watson==1.5.2
django-wishlist==1.0
djangorestframework==3.10.3
phonenumbers==8.10.23
Pillow==6.2.1
psycopg2==2.8.4
py-moneyed==0.8.0
python-utils==2.3.0
pytz==2019.2

